+++ src/com/example/exercises/singlylinkedlist/ListFactory.java
@@ -1,0 +1,11 @@
+package com.example.exercises.singlylinkedlist;
+
+public class ListFactory {
+
+	public static List<Integer> createSortedIntegerList() {
+		List<Integer> list = new List<>();
+		list.add(12);
+		list.add(5);
+		return list;
+	}
+}
